[pytest] Convert control plane assisted warm-reboot test (wr_arp) (#1572)
* [pytest] Convert control plane assisted warm-reboot test (wr_arp)

This PR converts Ansible control assisted warm-reboot (wr_arp) test
case to pytest.
