+++ tests/arp/files/ferret.conf.j2
@@ -1,0 +1,10 @@
+[program:ferret]
+command=/usr/bin/python /opt/ferret.py {{ ferret_args }}
+process_name=ferret
+stdout_logfile=/tmp/ferret.out.log
+stderr_logfile=/tmp/ferret.err.log
+redirect_stderr=false
+autostart=false
+autorestart=true
+startsecs=1
+numprocs=1
